CSL-245 move submit behaviour to common area
